added a CI pipeline for portfolio project

+++ edit_user.php
@@ -1,9 +1,14 @@
 <?php
 require('connect.php');
-require('authenticate.php');
+session_start();
+
+if (!isset($_SESSION['user_id'])) {
+    header("Location: index.html");
+    exit();
+}
 
 if (isset($_POST['delete'])) {
-    $id = $_POST['id'];
+    $id = filter_input(INPUT_POST, 'id', FILTER_SANITIZE_NUMBER_INT);
     $statement = $pdo->prepare("DELETE FROM users WHERE id = :id");
     $statement->bindValue(":id", $id);
     if ($statement->execute()) {
@@ -11,13 +16,27 @@ if (isset($_POST['delete'])) {
     }
 } else if ($_POST && isset($_POST['username']) && isset($_POST['role']) && isset($_POST['id'])) {
     $username = filter_input(INPUT_POST, 'username', FILTER_SANITIZE_FULL_SPECIAL_CHARS);
+    $email = filter_input(INPUT_POST, 'email', FILTER_SANITIZE_EMAIL);
     $role = filter_input(INPUT_POST, 'role', FILTER_SANITIZE_FULL_SPECIAL_CHARS);
+    $password = filter_input(INPUT_POST, 'password', FILTER_SANITIZE_SPECIAL_CHARS);
     $id = filter_input(INPUT_POST, 'id', FILTER_SANITIZE_NUMBER_INT);
 
-    $query = "UPDATE users SET username = :username, role = :role WHERE id = :id";
-    $statement = $pdo->prepare($query);
-    $statement->bindValue(':username', $username);        
+    if(!empty($password)) {
+        $hashed_password = password_hash($password, PASSWORD_DEFAULT);
+        $query = "UPDATE users SET username = :username, email = :email, role = :role, password = :password WHERE id = :id";
+        $statement = $pdo->prepare($query);
+        $statement->bindValue(':password', $hashed_password);
+        
+    } else {
+        $query = "UPDATE users SET username = :username, email = :email, role = :role WHERE id = :id";
+        $statement = $pdo->prepare($query);
+    }
+
+    
+    $statement->bindValue(':username', $username);
+    $statement->bindValue(':email', $email);
     $statement->bindValue(':role', $role);
+    
     $statement->bindValue(':id', $id, PDO::PARAM_INT);
 
     $statement->execute();
@@ -67,6 +86,7 @@ if (isset($_POST['delete'])) {
         
         <label for="username">Username</label>
         <input class="text-input" id="username" name="username" value="<?= $row['username'] ?>">
+        <input class="text-input" id="email" name="email" value="<?= $row['email'] ?>">
         <label for="role">Role</label>
         <select name="role" id="role">
           <option value="admin" 
@@ -78,6 +98,9 @@ if (isset($_POST['delete'])) {
             echo 'selected';
             }?>>Listener</option>
         </select>
+
+        <label for="password">Password</label>
+        <input type="password" class="text-input" id="password" name="password">
         
         <input class="button" type="submit" name="submit" value="submit">
         <input class="button" type="submit" name="delete" value="delete">
